Operate() function works recursively
The operate() function takes the users last input number, a running
total, and the function name of the button pressed to compute the
correct answer and update a global variable for a running total.
Functions to send the right args to operate() when the
corresponding button is pressed still need to be written.