appendCurrentNumber() working
appendCurrentNumber() works well. It takes an integer parameter and
converts it to and from a string to add it to the end of our
growing currentNumber global var.

This is done to make it so that our var can have numbers added to
the end of it as the user presses number keys. It also can accept
if the var is undefined because the user has pressed no keys yet.