Plus button adds
Plus button now successfully adds together numbers that are typed
into the display when the + button is pressed. Originally the
answer was returning NaN, but that's because I was clearing the
currentNumber before I was calling the opperate function. Once I
switched the order of these operations, the + function worked
just fine.