Add several opperations and write equals
Now plus() successfully adds together several integers successfully
and accurately. When the + button is pressed the function totals
together the last two expressions. Initializing currentAnswer =
currentNumber if it is empty makes this work.

Wrote an equals() function to clear screen, finish last expression,
and display the current total on the screen.

Next will write similar functionality for the remaining opperators.